[FIX/#22] Pose 프래그먼트로 변환
