+++ presentation/src/main/java/kr/genti/presentation/main/create/CreateViewModel.kt
@@ -17,6 +17,9 @@ class CreateViewModel
         val script = MutableLiveData<String>()
         val isWritten = MutableLiveData(false)
         var plusImage: Uri = Uri.EMPTY
+        val selectedAngle = MutableLiveData<Int>(-1)
+        val selectedFrame = MutableLiveData<Int>(-1)
+        val isSelected = MutableLiveData(false)
 
         private val _currentPercent = MutableStateFlow<Int>(33)
         val currentPercent: StateFlow<Int> = _currentPercent
@@ -27,5 +30,19 @@ class CreateViewModel
 
         fun checkWritten() {
             isWritten.value = script.value?.isNotEmpty()
+        }
+
+        fun selectAngle(itemId: Int) {
+            selectedAngle.value = itemId
+            checkSelected()
+        }
+
+        fun selectFrame(itemId: Int) {
+            selectedFrame.value = itemId
+            checkSelected()
+        }
+
+        private fun checkSelected() {
+            isSelected.value = selectedAngle.value != -1 && selectedFrame.value != -1
         }
     }
