+++ presentation/src/main/java/kr/genti/presentation/main/create/PoseFragment.kt
@@ -3,8 +3,10 @@ package kr.genti.presentation.main.create
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.activityViewModels
+import androidx.navigation.fragment.findNavController
 import dagger.hilt.android.AndroidEntryPoint
 import kr.genti.core.base.BaseFragment
+import kr.genti.core.extension.setOnSingleClickListener
 import kr.genti.presentation.R
 import kr.genti.presentation.databinding.FragmentPoseBinding
 
@@ -17,5 +19,19 @@ class PoseFragment() : BaseFragment<FragmentPoseBinding>(R.layout.fragment_pose)
         savedInstanceState: Bundle?,
     ) {
         super.onViewCreated(view, savedInstanceState)
+
+        initView()
+        initNextBtnListener()
+    }
+
+    private fun initView() {
+        binding.vm = viewModel
+    }
+
+    private fun initNextBtnListener() {
+        binding.btnPoseNext.setOnSingleClickListener {
+            findNavController().navigate(R.id.action_pose_to_selfie)
+            viewModel.modCurrentPercent(34)
+        }
     }
 }
