+++ presentation/src/main/res/layout/fragment_pose.xml
@@ -8,7 +8,7 @@
 
         <variable
             name="vm"
-            type="kr.genti.presentation.select.pose.PoseViewModel" />
+            type="kr.genti.presentation.main.create.CreateViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
